requirements.txt
pyyaml # for yaml config usage